Move data to datum folder. Remove routes to minimize build
