+++ client/datum/indicators.json
