+++ client/src/components/map/controls/info.tsx
@@ -8,7 +8,7 @@ import { LuInfo } from "react-icons/lu";
 
 import { cn } from "@/lib/utils";
 
-import { Indicator } from "@/app/local-api/indicators/route";
+import { Indicator } from "@/types/indicator";
 
 import Info from "@/containers/info";
