+++ client/src/containers/indicators/chart/imagery-tile.tsx
@@ -7,7 +7,8 @@ import { formatPercentage } from "@/lib/formats";
 import { useQueryImageryTileId } from "@/lib/indicators";
 import { useLocationGeometry } from "@/lib/location";
 
-import { Indicator, ResourceImageryTile } from "@/app/local-api/indicators/route";
+import { Indicator, ResourceImageryTile } from "@/types/indicator";
+
 import { useSyncLocation } from "@/app/store";
 
 import { CardLoader } from "@/containers/card";
