+++ client/src/containers/indicators/custom/municipalities.tsx
@@ -7,7 +7,8 @@ import { formatNumber } from "@/lib/formats";
 import { useGetIndicators, useQueryFeatures, useQueryFeatureId } from "@/lib/indicators";
 import { useLocationGeometry } from "@/lib/location";
 
-import { Indicator, ResourceFeature } from "@/app/local-api/indicators/route";
+import { Indicator, ResourceFeature } from "@/types/indicator";
+
 import { useSyncLocation } from "@/app/store";
 
 import { CardLoader } from "@/containers/card";
