+++ client/src/containers/indicators/numeric/imagery.tsx
@@ -5,7 +5,8 @@ import { useLocale } from "next-intl";
 import { useGetIndicatorsId, useQueryImageryId } from "@/lib/indicators";
 import { useLocationGeometry } from "@/lib/location";
 
-import { Indicator, ResourceImagery } from "@/app/local-api/indicators/route";
+import { Indicator, ResourceImagery } from "@/types/indicator";
+
 import { useSyncLocation } from "@/app/store";
 
 import { CardLoader, CardWidgetNumber } from "@/containers/card";
