+++ client/src/containers/indicators/numeric/index.tsx
@@ -5,7 +5,8 @@ import { useLocale } from "next-intl";
 import { useGetIndicatorsId, useQueryFeatureId } from "@/lib/indicators";
 import { useLocationGeometry } from "@/lib/location";
 
-import { Indicator, ResourceFeature } from "@/app/local-api/indicators/route";
+import { Indicator, ResourceFeature } from "@/types/indicator";
+
 import { useSyncLocation } from "@/app/store";
 
 import { CardLoader, CardWidgetNumber } from "@/containers/card";
