+++ client/src/containers/indicators/table/index.tsx
@@ -6,7 +6,8 @@ import { formatNumber } from "@/lib/formats";
 import { useQueryFeatureId } from "@/lib/indicators";
 import { useLocationGeometry } from "@/lib/location";
 
-import { Indicator, ResourceFeature } from "@/app/local-api/indicators/route";
+import { Indicator, ResourceFeature } from "@/types/indicator";
+
 import { useSyncLocation } from "@/app/store";
 
 import { CardLoader } from "@/containers/card";
