+++ client/src/containers/info/index.tsx
@@ -2,7 +2,7 @@ import { useLocale } from "next-intl";
 
 import { useGetIndicatorsId } from "@/lib/indicators";
 
-import { Indicator } from "@/app/local-api/indicators/route";
+import { Indicator } from "@/types/indicator";
 
 import { Markdown } from "@/components/ui/markdown";
 import { ScrollArea } from "@/components/ui/scroll-area";
