+++ client/src/containers/info/raster.tsx
@@ -2,7 +2,7 @@ import { useLocale } from "next-intl";
 
 import { useGetIndicatorsId } from "@/lib/indicators";
 
-import { Indicator } from "@/app/local-api/indicators/route";
+import { Indicator } from "@/types/indicator";
 
 export default function InfoRaster({ id }: { id: Indicator["id"] }) {
   const locale = useLocale();
